.ci/requirements.txt
semantic-version==2.10.0
bump2version==1.0.1
flake8==5.0.4
PyYAML==6.0.1
requests==2.32.0
click==8.1.3
click-completion==0.5.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
tenacity==8.2.3
